requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
streamlit
jupyter
google-generativeai
python-dotenv
